refactor: 예약 서비스 테스트 리팩토링 및 도메인 중심 설계 반영
기존 예약 서비스 테스트를 분리하고 도메인 중심 설계를 반영하여 WaitingDomainService와 ReservationDomainService를 추가.
